acq: use bar2_read_dma in acq.
Not the best design: the PciDevice object is exposed to the acq
controller, in order to create the kernel memory region and then set up
the DMA transfer.

The kernel memory region can (should) likely be shared across cores from
the same FPGA device. We should probably have a wrapper around PciDevice
that holds a vector of core controllers and handles memory reservation,
DMA transfers and kernel memory, and others.