pcie: flush bar0 and bar4 writes.
Problems with bar0 haven't been observed, but it's an easy change and
shouldn't impact anything.

With bar4, when writing (at least) more than 512 dwords (32-bit ints),
there were issues with corruption (random values being written to the
hardware or some writes not being registered -- the former were only
observed with HALCS, the latter here) and wishbone timeouts (all reads
return 0xffffffff). This suggests an issue with some part of the data
pipeline on the hardware side not being able to deal with so many writes
at once. Luckily, PCIe reads cause the writes to be flushed [1] [2], so
we can use that to make sure each write is done before starting the next
one, making it impossible to send too many packets at once.

Interestingly, neither MFENCE nor SFENCE seemed to flush the writes, so
the same issues could be observed when using them. Thankfully, forcing a
read is more portable and actually recommended.

[1] https://www.kernel.org/doc/html/latest/PCI/pci.html#mmio-space-and-write-posting
[2] https://groups.google.com/g/comp.os.vxworks/c/34_d90EfFa0