fofb_processing: add module.
We don't implement array input in the frontend, just a single value to
populate the whole memory bank. We only read from the first bank for
now, as well, pending on extensions to the general decoder interface to
support channels.

The float-to-fixed conversion, especially the use of round(), was based
on [1].

[1] https://embeddedartistry.com/blog/2018/07/12/simple-fixed-point-conversion-in-c/