decoders: extend RegisterDecoder to also hold data.
Decoders were only used to print decoded to a FILE stream, which made
them less useful as classes that can be used to build more complex
applications (such as an IOC).

Since we already have code that prints strings and values decoded from
hardware, and this code can be easily converted to do something else, we
are using strings as the keys. This gives us an additional advantage
when developing the IOC using the asyn module, since it will be able to
simply loop through its parameters and use their names as keys, without
adding a new line of code for each parameter, to be read from the
decoder.

Since printing now happens in a function in an upper layer, we need a
printers member. This also makes it necessary to install the printer.h
header.

The conversion is in progress, so compatibility members are kept around
(namely, constructor without printers parameter and dummy decode()
impl), but these should be removed as soon as possible.