lamp: fix code in channel implementation.
We shouldn't use the insert() method from std::vector, because it
increases the size of the vector. It was used with the intent of making
sure the vector would be extended to cover at least that index, but the
intent was wrong. We know the core has 12 channels, we can simply use
that value directly. This error wasn't detected, because the observable
behavior was still "correct", it was just also leading to absurd
slowdowns (and memory consumption) after library users had been running
for a while.