decoders: perform conversion to boolean automatically.
Values that are actually boolean will be automatically converted to
their 0-1 range. This is done by inspecting their Printer, as
that should be a source of truth and avoids the need to manually perform
the conversion to boolean everywhere in the decoders. This was already
done implicitly when printing via the decoder, but now users of the
library can take advantage as well.

An alternative is creating a get_bit() function, similar to
insert_bit(). It should be done eventually, but the current commit fixes
things across the board without making a lot of code changes.