fofb_cc: add module.
The information for the RAM based registers was taken from [1] and [2].
For the HALCS part, it was necessary to remove the 4x multiplier it used
on the values from the FPGA, due to using a byte-based address instead
of using the value as an index into an array of 32-bit values. This had
to be done for the offset as well, which was shifted two bits to the
right in our implementation (0x3 << 8 instead of 0x3 << 10).

One remaining task for this module is to implement the TOA and RCB parts
with a better interface.

[1] https://github.com/lnls-dig/halcs/blob/a86ddcabc7300cb035b3fcdeebd5f84229066fd6/include/hw/wb_fofb_cc_regs.h
[2] https://github.com/lnls-dig/CommsCtrlFPGA/blob/56bf6e653ff04b7620f8d2737f0fa950f4b666db/rtl/fofb_cc_pkg/rtl/vhdl/fofb_cc_pkg.vhd