acq: improve result performance.
Depending on the atom width for a given channel and the requested data
type, it was possible for acquisitions to go through 2 temporary vectors
before actually being returned to the user. One example was with signed
16-bit integers, where the acquisition:

  1. copied unsigned 16-bit ints from the device
  2. copied them to an unsigned 32-bit int vector
  3. copied them (with a sign extension step in the middle) to a signed
     32-bit int vector
  4. returned the signed 32-bit int vector to the user

The performance hit was noticeable, which motivated this change.

This commit avoids making the code too complex by using std::conditional
to declare the (possibly) temporary vectors; this way we have a single
implementation of each function, even while dealing with different
signedness. If at all possible, we return the first vector we created,
which was used to read data from the device -- this means callers must
know the atom width we are working with (this has yet to be exposed);
otherwise, we perform a single copy operation. An advantage of this
design is that the sign extension is performed directly by the type
system, so we avoid having to use the util.h functions, which added
another layer of indirection (though their performance impact was not
measured).

The measured result was that an acquisition went from 1.8s to 1.3s.