acq: change result_async interface.
To simplify locking for client applications, we make it so neither
result_async() nor get_result() have to run under a lock. Since we need
to read parameters in order to start the acquisition,
start_acquisition() has been made public; that's the only function that
needs to be run under a lock for clients using the
result_async/get_result combo. Clients using result() directly won't get
this benefit.

This change required:

- new member variables that will only be accessed by start_acquisition()
  and get_result(), so threaded clients only have to care about synching
  those two functions, and locking the class is only required for
  start_acquisition()
- new idle state for acq_status, since start_acquisition() now takes
  part in the state machine
- change the result_async() API so clients can decide when to copy the
  result vector using get_result. This also simplifies code that uses
  the return value of result_async
- moving the call to write_config() into start_acquisition()

This commit also includes an extra conditional in result() to avoid
sleeping when it's not needed.