acq: add stop_acquisition function.
This function is made idempotent, so clients can call it whenever
necessary. It is now also called during initialization of the class, to
make the whole system simpler. So stop_acquisition() doesn't use
uninitialized memory, we zero initialize the register storage [1].

In addition, remove FIXME comment about checking for currently running
acquisitions. The controller class now has a fully consistent view of
the hardware, and there is no reason for it to require this check.

[1] https://stackoverflow.com/questions/808464/c-new-call-that-behaves-like-calloc