pcie: remove DMA function.
Our solution currently does not support DMA, so it makes no sense to
leave this in the repository. We can also remove the comment in the acq
module about it; furthermore, the DMA transfer doesn't fit in the
current state machine.

If we revert this commit in the future, the FIXME comment can be removed
by simply adding a mutex for DMA transfers.