decoders: add support for storing float values.
We use double for storage, since it covers a wider range, and 32 bits of
precision (the max we can have with our registers, though we are usually
limited to 25) wouldn't fit in a float's mantissa [1].

The choice for using wrapper functions with different names instead of
parametrized ones from the start was because it would be necessary to
add a <class> parameter for most function calls, otherwise the type
deduction would get confused. Furthermore, making the necessary changes
would touch a bunch of files at once, and we aren't sure that requiring
specifying the type for all add_* function is reasonable; double
parameters are much less common.

We should strive to make the data storage class members private, in
order to simplify any future changes.

[1] https://en.wikipedia.org/wiki/Single-precision_floating-point_format#IEEE_754_standard:_binary32