util: start reorganizing header structure.
Split struct definitions from internal function declarations, to avoid
having to instal internal headers. This is a step in minimizing the
amount of headers that have to be installed. We now include "printer.h"
in all module implementations, in order to simplify removing it in the
future as well.

We also made "sdb-defs.h" C-compatible, and stopped installing "defer.h"
unnecessarily.