fofb_processing: increase ABI major version.
The ABI major version was increased [1] because the controller changed
how it calculates the error. The divergence from usual convention hadn't
been spotted previously due to the old FOFB IOC having a bug in the
fixed point conversion that inverted the signal of the inverse response
matrix values written to hardware.

[1] https://github.com/lnls-dig/fofb-ctrl-gw/commit/6a65e1947a44995cedf283a1d9eb6ad60e342d6b