acq: fix sample alignment for triggered acquisitions.
The previous align_extend function would allow the amount of samples to
be 0, even though that's not allowed by the acquisition core for all but
one case, as seen in the HALCS source [1]. We do allow - and in fact
require - post_samples to be 0 for immediate acquisitions, even though
the acquisitions only happen "after" they are started.

Since the align_extend function is only used in the acq module, and it
was going to require even more custom behavior, we chose to move it into
the acq implementation.

[1] https://github.com/lnls-dig/halcs/blob/6dfed5b6f5f02278696fb05bb7276a9d66cdb3fa/src/sm_io/modules/acq/sm_io_acq_exp.c#L180