sysid: work around cheby bug.
The struct definition had the wrong padding, so reads and writes to the
bpm_pos_distort RAM using the struct were reading and writing into the
padding for the first 4 words. This was detected because all reads from
the padding returned 0.

Since we changed the header manually, we have a static assertion to
guarantee that our fix stays there, and, if we update the header, that
either the same change is made or that the bug in cheby has been fixed.