acq: reject sample sizes that aren't a power of 2.
An issue was observed with the alignment of samples in the acquisition
buffer when the sample size wasn't a power of 2. This has yet to be
investigated fully, and should be tested at least in a situation where
the size of the acquisition buffer is a multiple of the sample size. For
now, we simply reject any such cases.