acq: subclass exceptions in use.
Our goal is to have start_acquisition() return errors when it can't
start a given acquisition, so we need to propagate information to it.
For now, we want to keep encode_config() and write_config() returning
void, as a standard across modules, so exceptions are the way to
propagate these errors.

Since we are here, also move the channel_atom_width check to
get_internal_values() instead of doing it in get_result().