decoders: make *_data private.
In order to achieve this, we need to resize the channel_data vector
ourselves, so modules don't have to. It was checked that all modules
were setting number_of_channels before calling add_channel*(), so it was
possible to simply use that member variable.

In lamp::CoreV2, be explicit about the case where we don't skip TRIG_EN.