afc_timing: implement trigger configuration.
Besides extending encode_config, this required resizing the parameter
vector during initialization, adding a default value to the source
parameter, and resetting the count_reset flag after writing to hardware.

We decided to use a vector string options for the source, instead of an
unordered_map, like other modules, because the list is sequential, and
this allows the list to be used by library users.