decoders: create RegisterDecoderBase.
This base class implements common functionality we have/would like to
see in RegisterDecoder and RegisterController. We chose to declare it in
"decoders.h" because it's tied more tightly to the decoding aspect.

For now, we are only replacing the common bits of RegisterDecoder, and
RegisterController will be migrated separately.

We are also adding extra functionality to set_devinfo(), which now
checks that the device passed to it is compatible with it, in order to
avoid any logic errors; however, no decoder makes use of it yet - the
plan is to remove the constructor that doesn't take the device
information parameters once everything's been migrated. get_data() is
using the new check_devinfo_is_set() function as well.