modules: create helper function and macro for regs.
It's defined in RegisterDecoderBase, as we use the new set_read_dest()
function in both decoders and controllers, since we plan on using the
read() implementation to remove some code from the controllers. The lamp
code is a special case for now, because it doesn't use the same
definitions; we plan on cleaning up its complicated inheritance once we
change the controller interface to take an array of parameters instead
of a channel parameter.

CONTROLLER_REGS was defined mostly because zero initializing the regs
struct wasn't being done for all devices, since it requires initializing
the class with "new type()", which is easy to forget.