controllers: define default write_params().
This way, subclasses which simply write the register structure to
wishbone can have less code. On top of that, we added a
check_devinfo_is_set() call to the default implementation, which helps
us catch logic errors.

We are also defining encode_params() as a virtual function, so
write_params() can use the implementation from each subclass.