fofb_shaper_filt: add module.
The static assertion for the amount of coefficients was written to make
sure the sizeof expressions were correct, but it also works as a check
for if the ABI ever changes.

The discrepancy in the amount of coefficients for Core and Controller is
intentional: for Core, we have only the actually relevant coefficients;
for Controller, we support the highest amount, so library clients don't
need to implement logic based on num_biquads. However, we can loop only
through the valid coefficients in both cases, for a slight optimization.

When actually writing to hardware, only the valid coefficients can be
written: both the 3 unused coefficients and the unused biquads must be
skipped. This was a bug noticed during development of the library, and
might be fixed in the future [1].

[1] https://github.com/lnls-dig/fofb-ctrl-gw/pull/36