decoders: remove need for data_order_done.
To avoid having to deal with the logic necessary for more complicated
channel layouts, the insertion into the data order vectors can implement
common logic for checking for duplicates, for the first time it's run.

The order vectors had to be changed to string_view to allow for
comparisons with std::find to work.