util: fix list_of_keys for std::unordered_map.
While list_of_keys() was fixed to allow compilation, what it was doing
was creating a string based on beginning and end iterators (case 7 for
C++11 in [1]) based on the two unrelated pointers, which failed some of
the time with std::length_error exceptions, and other times by printing
the wrong thing due to copying null bytes, but was always undefined
behavior.

std::string::append() always works, makes the function more consistent,
and avoids relying on the underlying string being null-terminated.

[1] https://cplusplus.com/reference/string/string/string/

Fixes: eb971fbee050987a8f463ba5e55d120133fe6441