decoders: remove skip from add_*_data functions.
It is no longer necessary, since we can deal with undefined printers;
since we don't have separate key vectors anymore, it also isn't
implemented. This presents a slight regression in the lamp module, which
now always prints TRIG_EN, but the correct value will still be printed
(TRIG_EN=0 is the same as the functionality missing entirely).