pcie: support serial wishbone interface.
For the PCIe interface, the functions remain the same. For BAR4 writes,
that means bar4_write_v() runs a loop of bar4_write() calls, so
bar4_write() is the one that actually interfaces with the hardware.
However, for serial interfaces, bar4_write() calls bar4_write_v() for
just one word, so bar4_write_v() is the one that interfaces with the
hardware.

We also set the fserport to null when it isn't initialized by one of the
dev_open functions.

Co-authored-by: Augusto Fraga Giachero <afg@augustofg.net>