pcie: add open_dev_serial.
This function currently doesn't deal with opening the serial interface
in an unknown state, so care should be taken to reset it before
launching anything using this library. Implementing it in
open_dev_serial would be a matter of writing '\n' to the device and
reading back and discarding any resulting message from the device.

Co-authored-by: Augusto Fraga Giachero <afg@augustofg.net>