ci: add alpine builds.
The "ubuntu" job name should have been "debian", but since we enhanced
the job to use multiple images, it was simply renamed to
"build-and-test".

For the Alpine builds, g++ is a separate package from gcc, unlike on
Debian. To simplify the job definition, we simply always install g++ on
Alpine. This isn't an issue, because we correctly export CXX to meson
setup.

We build with Alpine 3.18 because that's our platform for deployment. We
also build with Alpine Edge to take advantage of warnings from newer
compilers.