decoders: use one map for general and channel data.
This allows us to avoid dealing with two different interfaces (map of
values and map of vector of values) when handling register field data. A
good example is the substitution of add_general_internal() and
add_channel_internal() for a single function, which implements all the
necessary logic for registering data into the map.

A consequence of the new design is that we can now handle devices for
which a given field exists for X channels, and another field exists for
Y channels, with X != Y, in a cleaner manner. That means
number_of_channels is now simply the maximum number of channels among
all fields, and doesn't matter otherwise. This change brings some
advantages: get_channel_data() calls will now error out if a field
doesn't exist for a given channel, even if it's inside the
number_of_channels range; and print() will only print fields for the
channels where they actually exist.

It was necessary to change the print() function to handle this new
design, but it was a small change.

Also rename RegisterDecoder::data to RegisterDecoder::pvt, to avoid the
data->data repetition and to be clearer about its purpose.

Benchmark comparison on my notebook (Intel i7-1185G7) (in ns):

                               | original | new |
|------------------------------|----------|-----|
| Set general values           | 200      | 215 |
| Get general values - integer | 58       | 74  |
| Get general values - double  | 72       | 82  |
| Set channel values           | 716      | 470 |
| Get channel values - integer | 71       | 77  |
| Get channel values - double  | 77       | 96  |

There was a small performance decrease for most benchmarks (biggest
being "Get general values - integer" with +~25% in execution time).
Unexpectedly, "Set channel values" had a big performance improvement.
Even so, the overall negative performance impact is considered
acceptable for the interface and maintainability benefits it brought.