decoders: add RegisterField.
As opposed to add_* functions which take int or double values directly,
the RegisterField variants of these functions, which take the structs
generated by rf_extract_value() and rf_get_bit(), make it possible to
store data about a register field's origin: address offset in register
map, bit mask, and sign.

To keep the interface consistent and avoid issues, the actual value of
the register is still stored the way it would be with the "normal" add_*
functions.

For now, we simply store that information; furthermore, double values,
which require conversion to and from fixed point, aren't handled yet.