util: change extract_value interface.
Instead of having to specify how to interpret the values in terms of
sign *and* bit-width, we can specify only the sign and use the known
bit-width of the binary mask. On top of that, the sign can be a function
parameter instead of a template one.

We also add checks for the validity of the masks: more than one bit is
set, and all bits are contiguous.

To avoid breaking all existing users of the extract_value templates in
modules, we keep the templates as a temporary compatibility measure, but
implement them as a call to the new function.