controllers: add RegisterDecoderController.
This is a controller superclass which uses the
write_general/write_channel methods from the module's decoder to avoid
the need to duplicate information across the decoder and encoder.

We can implement a default (and empty) encode_params() member function
here, since it's now possible that modules won't require any special
handling for encoding beyond the write_general() and write_channel()
functions.