util: add get_index function.
Uses the string search logic from clear_and_insert_index(), but simply
returns the index. Will be useful for modules using
RegisterDecoderController, for example, where we don't want to have a
parameter struct that takes an std::string, but still want to be able to
set a given enum by name (e.g. when using decode-reg).

Also change the interface to use a const reference to the value_list
vector to avoid potentially expensive copies.