decoders: add fixed point support to RegisterField.
Define rf_fixed2float function for converting an integer RegisterField
to fixed point, and change write_* interface to take decoders::data_type
so users can write doubles to the fixed point fields.

RegisterField::fixed_point_pos is set to 0 in the struct definition so
it isn't a required field when instantiating the struct.