decoders: implement get_generic_data.
This function is useful for library users which need to handle register
field values in a generic way, since it's a single call point for both
kinds of data and for general and channel data.

Also implement get_general_data() and get_channel_data() as calls to it,
which simplifies the code and error handling, and avoids the need to
specify templates while at the same time improving compile time type
checks.