util: add si57x class.
This class implements the same overall algorithm as
modules/afc_timing.cc, but in a way which can be reused by other
modules.

There is a small change for better precision: we use llrint() instead of
automatic truncation when assigning from a double expression into an
integer variable. llrint() has been chosen in place of llround() since
it implements half-even rounding, like python's round(), which is used
in our reference implementation of this algorithm [1].

[1] https://github.com/lnls-dig/afc-gw/blob/eb03f2a6b4653d1831e80d305d0977ba4d6597b6/scripts/si57-clk-calc.py