ad9510: add module.
HALCS used the address offset between the FMC_ACTIVE_CLK and SPI
registers to identify which SPI core was connected to the AD9510 IC
belonging to that FMC active clock (FMC_ACTIVE_CLK_AD9510_SPI_OFFS in
[1]). Since we don't directly instantiate the object responsible for the
SPI core in the active clock object like HALCS does [2] -- instead, they
are identified by their shared index value -- it is necessary to
implement some way of identifying the SPI core. We decided to go with
the module order, which is less likely to change than register map sizes
and offsets.

[1] https://github.com/lnls-dig/halcs/blob/8b8d49253a7dd9a403cb1e59d382f4c0738627b2/include/mem_layout_common.h
[2] https://github.com/lnls-dig/halcs/blob/8b8d49253a7dd9a403cb1e59d382f4c0738627b2/src/sm_io/modules/fmc_active_clk/sm_io_fmc_active_clk_core.c#L57