decode-reg: error out when build_info is empty.
In an inconsistent link state, it is occasionally possible to open the
resource file for a given board, even though it does not return any
valid information when the mapped BARs are read. In this case, there is
no SDB filesystem to navigate, or synthesis information to be reported,
which will result in build_info being empty, and either print nothing,
or cause a segmentation fault in quiet mode.

Therefore, we add an actual error message and return an error value.

Co-authored-by: Érico Nogueira <erico.rolim@lnls.br>