decoders: don't use boost::hash.
This avoids depending on boost for anything, which simplifies the list
of dependencies.

Instead, we inline their hash_combine implementation, including the
magic constants, even if it isn't the best possible hash combinator [1].
The hasher class organization was taken from [2].

Since the hash of std::optional<unsigned> should be the same as the hash
of unsigned when it contains a value [3], and on libstdc++ and libc++
that hash is simply the identity function [4], it felt more appropriate
to use hash_u (the second member of the pair) as the seed, so it gets
some additional mixing.

[1] https://stackoverflow.com/q/35985960/9512804
[2] https://stackoverflow.com/a/20602159/9512804
[3] https://en.cppreference.com/w/cpp/utility/optional/hash
[4] https://tessil.github.io/2016/08/29/benchmark-hopscotch-map.html