util: clean up Printer constructor interface.
The interface for setting custom strings for PrinterType::boolean wasn't
very intuitive. Taking the option for "false" (0) first, and then for
"true" (1), is less error prone, because it isn't necessary to remember
that this function is a special case, it simply follows numerical order.

Instead of always requiring a PrinterType argument, it is now
automatically set by the constructor depending on the Constructor
version that has been called; only the fields using the built-in
messages pass PrinterType now. Therefore, user code doesn't have to use
PrinterType::custom_function anymore, so it was made internal (only in
name).

Removing usage of PrinterType::custom_function was done by hand, while
changing the order of custom strings and removing usage of
PrinterType::boolean was done using the sed(1) command below, except for
a special case in fmcpico1m_4ch.

    sed -i 's/PrinterType::boolean, "\(.*\)", "\(.*\)"/"\2", "\1"/' \
        $(git grep -l PrinterType::boolean)