util: add more information to SDB printing.
This required passing more information to print_sdb(), which we have now
made into an internal function, to avoid having to expose the internals
of the sdbfs library. This also gave us the opportunity to make some
changes to the read_sdb() interface, which is now a single function,
without any overloads.