templates: add module template.
This makes it simpler to add the boilerplate for new modules, instead of
having to copy an existing file and remember to change all the required
places.

The script doesn't handle adding the modules to meson.build yet, but
that might be an interesting future direction.

We add the templates to the build to ensure they don't become outdated.
In order to be able to build them without any hacks, mod_template.cc had
to use a simple include for "mod_template.h", so add_module.sh has to
fix the include path.