clang-tidy: add config.
The test groups that were added were the ones considered relevant for
this project. Test groups used to enforce code style for specific
projects and platforms were omitted, as was cppcoreguidelines, because
it's harder to apply to a codebase that wasn't using it from the start.

The specific tests from the groups in use that had to be removed are
explained below:

- bugprone-empty-catch: too eager, recommends longer code using find()
  member function;
- concurrency-mt-unsafe: flags usage of exit(3) and strerror(3), which
  is basically noise. exit(3) has to be called by someone, and
  strerror(3) is thread-safe on the common case with locale set up and
  valid error values;
- modernize-use-trailing-return-type: though more modern, this function
  declaration style isn't widespread, and would conflict with the style
  used in other projects this one interacts with. It would also increase
  the barrier to entry for this project;
- performance-avoid-endl: we want to be able to use endl;
- readability-braces-around-statements: braces are skipped for single
  statements, which should always be indented (this will be guaranteed
  by clang-format in the future);
- readability-identifier-length: short names for iterator variables and
  throwaway values, as well as known shorthand (f for FILEs, v for
  verbose) should be used;
- readability-implicit-bool-conversion: treating pointers as bools is a
  common idiom;
- readability-named-parameter: parameters are named in implementation
  files as a design choice.

We want clang-tidy to verify our own headers, which are under
include/modules/ and utils/.

We also want clang-tidy-diff, when used, to exit with an error.