ci: add static analysis workflow.
Use a matrix with jobs in order to parallelize running different tools.

The check for GITHUB_HEAD_REF is used so the same workflow works for PRs
and for branches.

We ignore some files with cppcheck due to issues with Catch2. An example
of an error generated by cppcheck with Catch2:

  util/tests/bits-test.cc:36:5: error: There is an unknown macro here
  somewhere. Configuration is required. If _catch_sr is a macro then
  please configure it. [unknownMacro]
    CHECK_THROWS_AS(clear_and_insert(reg, 1000U, range_mask), std::runtime_error);

The issue causing git-config(1) to be necessary is tracked in [1].

Meson supports a clang-tidy target of its own, but it's not really
usable for our purposes [2].

[1] https://github.com/actions/checkout/issues/1169
[2] https://github.com/mesonbuild/meson/issues/2383