ci: add builds with all supported sanitizers.
This way, tests will be run with sanitizers as well. Since including a
sanitizer in static binaries isn't recommended (doesn't work with LLVM,
with GCC requires a special flag), we include the sanitizer flags in the
same matrix group as linking flags.

The "memory" sanitizer has been omitted for now, since GCC doesn't
support it.