util: avoid memory leak in copy-test.
dummy_dev_open was using malloc, so, when the test was finished,
AddressSanitizer detected a leak from the pcie_bars struct. In order to
allow us to take advantage of the existing dev_close implementation for
this test, we switched dummy_dev_open to use mmap instead of malloc.

Using mmap means the mappings are no longer tracked by malloc, so
avoiding memory leak reports by calling dev_close is actually no longer
necessary (yes, the situation is a bit circular). But it's good practice
to close all handles, since it can make the program state easier to
analyse.

Reported-by: LeakSanitizer