fofb_cc: work around cppcheck warning.
Without this change, cppcheck would accuse that target_reg was
uninitialized when used, even though a reference had been captured, and
the lambda was only called after it was initialized.

We can reorder the code so this false-positive isn't triggered, and
without accidentally suppresing any warnings from other tools (as might
have happened if we had initialized target_reg to nullptr).