sdbfs: use subproject instead of submodule.
The motivation for using a subproject is to speed up cloning when
building uhal, since submodules are usually cloned recursively. It also
improves consistency for third party dependencies.

We define our own wrap file and meson.build for the fpga-config-space
project, though we only build sdbfs. We also transition to the LNLS fork
of this repository, in order to avoid depending on the OHWR GitLab
instance.

The LNLS fork no longer requires -Wno-language-extension-token, so it's
been removed from the project definition. However, it should be noted
that the justification for commit ec083dd (sdbfs: disable
-Wlanguage-extension-token warning., 2023-07-26) was wrong, as we've
always had push access to the fpga-config-space repository, and could
just as well have already removed this requirement there.