ordered-map: use subproject instead of submodule.
The motivation for using a subproject is to speed up cloning when
building uhal, since submodules are usually cloned recursively. It also
improves consistency for third party dependencies.

Since ordered-map is a header-only library, we were previously using its
include directories variable. Since we are now using it as a subproject,
it needs to be used as a dependency.

This also allows us to fully remove the third_party subdirectory.