ci: add sanitizer workflow.
Using sanitizers on Alpine doesn't seem to be supported properly, so we
are sticking with Debian (glibc) only.

Enabling sanitizers can change some compiler passes, possibly generating
false-positive warnings [1]. Warnings are already checked by other
workflows, so we omit -Werror from this one.

  Note that sanitizers tend to increase the rate of false positive
  warnings, most notably those around -Wmaybe-uninitialized. We
  recommend against combining -Werror and [the use of] sanitizers.

In order to improve error messages, we build with debug information.

We didn't (explicitly) use all of the available sanitizers, such as:

- MemorySanitizer isn't supported by GCC and requires that all code be
  instrumented, including dependencies [2]
- LeakSanitizer is already run as a last phase of AddressSanitizer [3]

ThreadSanitizer is used even though it isn't relevant for now, since we
don't have any multi-threaded tests yet. However, we can add them in the
future and make use of it.

[1] https://gcc.gnu.org/onlinedocs/gcc/Instrumentation-Options.html#index-fsanitize_003dbuiltin
[2] https://github.com/google/sanitizers/wiki/MemorySanitizer#using-instrumented-libraries
[3] https://github.com/google/sanitizers/wiki/AddressSanitizerLeakSanitizer